spark/requirements.txt
# Apache Spark and Delta Lake dependencies
pyspark==3.5.0
delta-spark==3.0.0

# Kafka integration
kafka-python==2.0.2

# Data processing and utilities
pandas==2.1.4
numpy==1.24.3
pyarrow==14.0.1

# AWS/S3 compatibility for MinIO
boto3==1.34.0
s3fs==2023.12.2

# Configuration and environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-spark==0.6.0

# Development tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1